feat(dasclaw_runtime): port SecretsStore trait + secret types to shared crate (F3.2 phase 2 PR 4a)
将 ironclaw 内 secrets 模块的纯类型层 + 抽象 trait 迁移到 dasclaw_runtime::secrets:

新增 crates/dasclaw_runtime/src/secrets/:
- types.rs: Secret / SecretRef / DecryptedSecret / SecretError /
  CreateSecretParams / CredentialLocation / CredentialMapping (~500 行,
  纯数据,无 DB / OS 依赖)
- store.rs: SecretsStore trait + 一个内存 mock 用于契约测试 (~200 行)
- mod.rs: re-exports

ironclaw 通过 re-export 透明转发, 不修改任何调用方 (~20+ 处导入路径全部保留):
- desktop-client/ironclaw/src/secrets/types.rs -> pub use dasclaw_runtime::secrets::types::*
- desktop-client/ironclaw/src/secrets/store.rs 删除 trait 定义, 改为 pub use,
  保留所有具体 impl (Postgres / LibSQL / InMemory / agent_provider / crypto / keychain)
  —— 后者留给 sub-PR 4b (crypto + keychain) / 4c (具体后端) 处理。

测试: 46/46 通过 (本 PR 新增 27 个 req_dasclaw_runtime_secrets_*, 含 trait
契约测试覆盖 fail-safe 默认拒绝 is_accessible)。

闸门: cargo check (runtime + dasclaw) / nextest 46/46 / clippy -D warnings /
fmt / check_no_panics / check_no_new_ironclaw_literal — 全绿。

Closes #641 (F3.2 phase 2 PR 4a of 6)
