+++ crates/dasclaw_runtime/Cargo.toml
@@ -10,6 +10,10 @@ dasclaw_tool = { path = "../dasclaw_tool" }
 thiserror = "1"
 serde = { version = "1", features = ["derive"] }
 chrono = { version = "0.4", features = ["serde"] }
+secrecy = { version = "0.10", features = ["serde"] }
+uuid = { version = "1", features = ["v4", "serde"] }
+async-trait = "0.1"
 
 [dev-dependencies]
 serde_json = "1"
+tokio = { version = "1", features = ["macros", "rt"] }
